bug: correct the layer calculation
The first layers (0 and 1) were being populated incorrectly. Modules with no
external requirements were being assigned a layer below a module with any
requirements even if those requirements were all outside the set of modules
being considered. So a module requiring just the standard library would be in
a layer above a module with no external requirements. The fix was to check
that the required module's go.mod file had itself been read (so it had a
non-nil Loc)